Revert teams validation from model to form
This is because apparently it is a huge pain in the ass to get Django to validate
a many to many field in the model. This is because the models needs to be saved
before the connection can be made.

Both sides of the relationship need to be saved separately before the
relationship can be created.

So now it seems the easiest to just leave it in the form.