Use default formset rendering
I guess when I was messing with the views, I changes the form rendering

This had now let to the form not working because the expected hidden fields for game and pick
where not present.

Now these fields are automatically included again, because the formset is rendering itself.

The question remains, wether I really need the game and the pick in there.