Remove pick from game pick form
Pick and Game Pick form are shown in and created in the same view.
Therefore, it can not be known before hand what the pick value of the
game pick is going to be. It therefore makes no sense to have that value on the
form.

This commit removes the useless form field.