Show game name as field label in game pick form
Because the form is so small, it should be obvious that you are going to pick the
winning team. The label "picked team" does not really make any sense in that context.

Now the label shows the game name.

This is especially more useful since these forms will be shown as a set.
Once styling comes into play, it might make sense to generate the form manually,
but for now, this is going to be enough.