Replace GameFactory class method with hook
Before, I had only implemented a class method to allow generation of the many
to many relationship from Game to Team.

That was not ideal, because that prevents the Game factory being used as a
SubFactory in other Factories.

Now I am making use of the Facotry Boy `post_generation` hook. This is the
documented way of creating simple many to many relationships like the one I
have.

See also:
https://factoryboy.readthedocs.io/en/stable/recipes.html#simple-many-to-many-relationship