Move Tally view to picks app
The model that is being listed is the PickSheet. So it makes sense to
have the view that lists PickSheets in a particular order inthe picks
app. This is in line with the other changes that moved the forms to the
app that defines the models the form is based on.

The fact that all these thing requrie a pot as well does not mean they
need to be in the pot app. Just like the models depend on the pot, so
can the forms and the views.