Move set_results view to games app
This one is a bit less clear because the tiebreaker is associated with
the pot so far. But, I can see how that might change and a abstract
model like a "game set" will be introduced to hold all the games of a
pot. I could imagine that being helpful for things like different
tiebreaker evaluation etc.